Finish Game
I was too focus and had too much fun that I forgot to documented per steps in my GitHub. The trigger was when I need to make Finite State Machine for enemy in the simplest form without wasting too much time but still robust and have many if statements in the update. The concept of FSM is making action into a seperated form, I use delegate Action to realize that very concept. As a result, the logic for each Action has less coupling and in the update it's not bloated with if statements.